pyproject.toml
[project]
name = "exam-management-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cookiecutter>=2.6.0",
    "cryptography>=46.0.5",
    "django>=6.0.2",
    "django-tailwind[reload]>=4.4.2",
    "pillow>=12.1.1",
    "pyotp>=2.9.0",
    "python-dotenv>=1.2.1",
    "qrcode>=8.2",
    "uvicorn>=0.40.0",
    "whitenoise>=6.11.0",
]
